fix: split header.js on three separate files
