+++ english-for-kids/src/index.html
@@ -10,7 +10,6 @@
 </head>
 
 <body>
-    <script src="./index.js"></script>
 </body>
