+++ english-for-kids/src/index.js
@@ -1,47 +1,4 @@
 import './main.css';
-
-// import { header } from './scripts/header';
-// import {mainDOM} from "./scripts/spa";
-
-import { switchToStateFromURLHash } from './scripts/spa';
-import hug from './assets/imgs/hug.jpg';
+import { switchToStateFromURLHash } from './scripts/spa/spa';
 
 switchToStateFromURLHash();
-
-export const img = new Image();
-img.src = hug;
-
-// let header = document.createElement("header");
-// header.classList.add("header");
-// let header__container = document.createElement("div");
-// header__container.classList.add("container");
-
-// let header__burger = document.createElement("div");
-// header__burger.classList.add("header__burger");
-// header__burger.innerHTML = `
-//     <svg viewBox="0 0 80 60" width="40" height="40">
-//         <line x1="4" y1="0" x2="76" y2="0"></line>
-//         <line x1="4" y1="25" x2="76" y2="25"></line>
-//         <line x1="4" y1="25" x2="76" y2="25"></line>
-//         <line x1="4" y1="50" x2="76" y2="50"></line>
-//     </svg>
-// `;
-
-// let h1 = document.createElement("h1");
-// h1.textContent = "English for kids";
-// document.body.appendChild(header);
-
-// let header__switch = document.createElement("div");
-// header__switch.classList.add("header__switch");
-// header__switch.classList.add("checked");
-// header__switch.innerHTML = `
-//     <span>train</span>
-//     <div class="header__switch-round"></div>
-//     <span>play</span>
-// `;
-
-// header__container.appendChild(header__burger);
-// header__container.appendChild(h1);
-// header__container.appendChild(header__switch);
-// header.appendChild(header__container);
-// document.body.appendChild(header);
