+++ english-for-kids/src/scripts/header/model.js
@@ -1,0 +1,49 @@
+// !!! MENU MODEL !!!
+// !!! MENU MODEL !!!
+// !!! MENU MODEL !!!
+
+const menuArray = [
+  {
+    innerText: 'Main page',
+    id: 'main_page_btn',
+  },
+  {
+    innerText: 'Action (set A)',
+    id: 'actiona_page_btn',
+  },
+  {
+    innerText: 'Action (set B)',
+    id: 'actionb_page_btn',
+  },
+  {
+    innerText: 'Action (set C)',
+    id: 'actionc_page_btn',
+  },
+  {
+    innerText: 'Adjective',
+    id: 'adjective_page_btn',
+  },
+  {
+    innerText: 'Animal (set A)',
+    id: 'animala_page_btn',
+  },
+  {
+    innerText: 'Animal (set B)',
+    id: 'animalb_page_btn',
+  },
+  {
+    innerText: 'Clothes',
+    id: 'clothes_page_btn',
+  },
+  {
+    innerText: 'Emotion',
+    id: 'emotion_page_btn',
+  },
+  {
+    innerText: 'Statistics',
+    id: 'statistics_page_btn',
+  },
+
+];
+
+export { menuArray };
